% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnovaDF.R
\name{AnovaDF}
\alias{AnovaDF}
\title{Tabla Anova de Diseño Factorial}
\usage{
AnovaDF(Datos, a, b, r, alfa)
}
\arguments{
\item{Datos}{(dataframe) Tabla que contiene los datos a analizar, donde se
deberán tener 5 columnas, la primera que indique los niveles del Factor A,
la segunda los niveles del Factor B, la tercera los Tratamientos, otra con las
Réplicas y la última que contenga las Respuestas.}

\item{a}{(integer) Número de niveles del Factor A.}

\item{b}{(integer) Número de niveles del Factor B.}

\item{r}{(integer) Número de réplicas de cada tratamiento.}

\item{alfa}{(numeric) Indica la significancia que se tomará para el diseño experimental.}
}
\value{
Devuelve la Tabla Anova que contendrá los grados de libertad, la Suma de
Cuadrados, los Cuadrados Medios, los valores de F calculada y de F de tablas.
}
\description{
Analiza los datos obtenidos de un diseño factorial, es decir, nos ayuda a
analizar los resultados que obtenemos de un Diseño Factorial 2x2, mediante una
Tabla Anova, para poder concluir si se rechaza o no la Hipótesis Nula.
}
\examples{
\dontrun{
##Limpiar la memoria de R
rm(list = ls())
#----------------------------------------------------------------------------
#EJEMPLO 1

##Se establece la ruta del archivo y se fija
ruta <- "C:.../Datos"
setwd(ruta)

##Importar la base de datos a un Dataframe
datos <- read.csv(file = "Datos.csv")
df <- data.frame(datos)
print(df)

##Se carga la librería
library(DisFac)

##Se ejecuta la función para obtener la Tabla Anova
TablaAnovaDF(Datos = "df", a = "2", b = "2", r = "4", alfa = "0.05")
}
}
