% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MediasFactor.R
\name{MediasFactor}
\alias{MediasFactor}
\title{Medias por nivel de cada Factor}
\usage{
MediasFactor(f, o)
}
\arguments{
\item{f}{(integer) Indica el factor del cual se pretende calcular las medias, 1 es para el Factor A, y 2 para el Factor B.}

\item{o}{(integer) Indica el número de observaciones para cada factor.}
}
\value{
Devuelve una lista con las medias por cada nivel del factor.
}
\description{
Hace el cálculo de Medias para cada nivel de cada factor del diseño Factorial.
}
\examples{
\dontrun{
#----------------------------------------------------------------------------
#EJEMPLO
#Si se rechaza la Hipótesis Nula mediante los valores de F calculada y F de tablas obtenidas en la tabla
#Anova, se hace un calculo de Medias por nivel del Factor.

##Se ejecuta la función para calcular la Medias por Factor
Medias(f = "1", o = "8")
}
}
