% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PruebaDms.R
\name{pDms}
\alias{pDms}
\title{Prueba de Diferencia Mínima Significativa}
\usage{
pDms(alfa, r, glE, CME)
}
\arguments{
\item{alfa}{(numeric) Indica la significancia que se tomará para la prube DMS.}

\item{r}{(numeric) Indica el número de réplicas de cada tratamiento en el experimento.}

\item{glE}{(numeric) Indica los grados de libertad del Error, que obtendremos de la tabla Anova.}

\item{CME}{(numeric) Indica el Cuadrado Medio del Error que obtendremos de la Tabla Anova.}
}
\value{
Devuelve dos tablas, la primera que contiene las medias por tratamientos, y la segunda que contendrá las diferencias entre las medias por tratamientos, y señala cuales son mayores respecto al valor del DMS.
}
\description{
Hace una Prueba DMS, para que mediante las medias podamos obtener el mejor tratamiento o combinación de los factores.
}
\examples{
\dontrun{
#----------------------------------------------------------------------------
#EJEMPLO

##Se ejecuta la función para realizar la Prueba DMS
PruebaDms (alfa = "0.05",r = "4", glE = "12", CME = "24.31")
}
}
